+++ seaice3p/params/dimensional/__init__.py
@@ -2,6 +2,7 @@ from .ocean_forcing import (
     DimensionalBRW09OceanForcing,
     DimensionalFixedHeatFluxOceanForcing,
     DimensionalFixedTempOceanForcing,
+    DimensionalMonthlyHeatFluxOceanForcing,
 )
 from .dimensional import (
     DimensionalParams,
@@ -23,6 +24,7 @@ from .forcing import (
     DimensionalTurbulentFlux,
     DimensionalRobinForcing,
     DimensionalERA5Forcing,
+    ERA5FileKeys,
 )
 from .initial_conditions import (
     BRW09InitialConditions,